feat: to_html for new node types

+++ src/build/mod.rs
@@ -1,5 +1,8 @@
 use build_md_file::build_md_file;
-use std::{fs::{self, remove_dir, remove_dir_all}, vec};
+use std::{
+    fs::{self, remove_dir_all},
+    vec,
+};
 use tera::Tera;
 use walk_content_dir::walk_content_dir;
 use walk_static_dir::{copy_static_file, walk_static_dir};
@@ -13,7 +16,7 @@ mod walk_content_dir;
 mod walk_static_dir;
 
 /// Run through md files in content and generate html from them!
-pub fn build() {
+pub fn build(empty_out_dir: bool) {
     let proj_dir = get_project_dir();
     let mut tera = match Tera::new("../templates/**/*.html") {
         Ok(t) => t,
@@ -23,7 +26,9 @@ pub fn build() {
     };
     tera.autoescape_on(vec![]);
     // clear out stale files
-    remove_dir_all("public").unwrap();
+    if empty_out_dir {
+        remove_dir_all("public").unwrap();
+    }
     fs::create_dir("public").unwrap();
     walk_content_dir(&proj_dir.join("content"), &tera, build_md_file).unwrap();
     walk_static_dir(&std::path::Path::new("../static"), copy_static_file).unwrap();
